+++ backend/dataall/modules/worksheets/services/__init__.py
@@ -2,6 +2,6 @@
 Contains the code needed for service layer.
 The service layer is a layer where all business logic is aggregated
 """
-from dataall.modules.worksheets.services import services, permissions
+from dataall.modules.worksheets.services import worksheet_services, worksheet_permissions
 
-__all__ = ["services", "permissions"]
+__all__ = ["worksheet_services", "worksheet_permissions"]
